requirements.txt
azure-storage-blob
python-multipart
uvicorn
fastapi~=0.100.0
